Title: 세 수의 합과 평균 구하기, Time: 131ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240205/세 수의 합과 평균 구하기/sum-and-mean-of-three-numbers.java
@@ -8,6 +8,6 @@ public class Main {
         int b = sc.nextInt();
         int c = sc.nextInt();
 
-        System.out.printf("%d\n%.0f", a + b + c, (double) (a + b + c) / 3);
+        System.out.printf("%d\n%d", a + b + c, (a + b + c) / 3);
     }
 }
